requirements.txt
requests == 2.22.0
setuptools == 40.8.0
PyYAML == 5.1.2
pymongo == 3.9.0